+++ models/json-persistance/crud.js
@@ -1,18 +1,18 @@
 const { readFile, writeFile } = require("fs/promises");
 const path = require("path");
-const getCurrentTime = require("./helpers/get-current-time");
-const Task = require("./data/models/task");
+const getCurrentTime = require("../../helpers/get-current-time");
+const Task = require("./task");
 const { writeFileSync } = require("fs");
 
 function init() {
-  writeFileSync(path.join(__dirname, "./data/json-db.json"), "[]");
+  writeFileSync(path.join(__dirname, "../json-db.json"), "[]");
   console.log(`Database connection established`);
 }
 
 // Returns database as object 
 async function findAll() {
   try {
-    const db = await readFile(path.join(__dirname, "./data/json-db.json"));
+    const db = await readFile(path.join(__dirname, "../json-db.json"));
     return JSON.parse(db);
   } catch (err) {
     console.log(err.message);
@@ -50,6 +50,7 @@ async function update(id, status) {
     dbAsArray[userPosition].status = status;
     dbAsArray[userPosition].endedAt = getCurrentTime();
     await write(dbAsArray);
+    return dbAsArray;
   } catch (err) {
     console.log(err.message);
   }
@@ -61,6 +62,7 @@ async function deleteOne(id) {
     const userPosition = dbAsArray.findIndex((user) => user.id === id);
     dbAsArray.splice(userPosition, 1);
     await write(dbAsArray);
+    return dbAsArray;
   } catch (err) {
     console.log(err.message);
   }
@@ -68,9 +70,10 @@ async function deleteOne(id) {
 
 async function write(db) {
   await writeFile(
-    path.join(__dirname, "./data/json-db.json"),
+    path.join(__dirname, "../json-db.json"),
     JSON.stringify(db)
   );
 }
+
 
 module.exports = { init, create, findOne, findAll, update, deleteOne };
